+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Match memory regions to pack file specified core names.
 - `probe_rs_target::chip::Chip` has a new field `pack_file_release` which is populated by `target-gen`.(#1259)
 - Benchmarking code moved from an example to `probe-rs-cli` subcommand (#1296).
+- Replace `log` crate, with `tracing` in `probe-rs-debugger` executable, and in the `rtt` library. ()
 
 ### Fixed